fix(dashboard): satisfy eslint on Advanced settings deeplink
Move the URL helper out of the component module and remount the
collapsed group from the route instead of setting open state in an
effect.
