+++ frontend/src/features/settings/advanced-settings-deeplink.ts
@@ -1,0 +1,7 @@
+export function shouldExpandAdvancedSettings(search: string, hash: string): boolean {
+  const query = search.startsWith("?") ? search.slice(1) : search;
+  if (new URLSearchParams(query).get("advanced") === "1") {
+    return true;
+  }
+  return hash === "#firewall";
+}
